Modify Handler to call previously defined handlers
Both set_exception_handler and set_error_handler return the values to
which they were previously set. Be respectful and call these handlers
as it may introduce unexpected behavior in the user's code and also
provides no way to attach an error handler besides manually wrapping
the bugsnag library.

PHPUnit's default behavior is to translate PHP errors into exceptions
and raise those. Rely on this assumption in our tests of the error
handler.

Ref: https://phpunit.de/manual/current/en/writing-tests-for-phpunit.html