$callPrevious should not be an optional parameter
Since these are new methods expected to be invoked primarily by register
or registerWithPrevious, make the boolean flag a required parameter.

A user might wish to only register certain handlers themselves, so it
makes sense to leave these public, but the behavior is different
depending on the flag. The caller should be forced to make an explicit
decision.