Call previously defined exception and error handlers (#395)
* Add test cases to show errant behavior

PHPUnit's default behavior is to translate PHP errors into exceptions
and raise those. Rely on this assumption in testErrorHandler.

Ref: https://phpunit.de/manual/current/en/writing-tests-for-phpunit.html

* Modify Handler to call previously defined handlers

Both set_exception_handler and set_error_handler return the values to
which they were previously set. Be respectful and call these handlers
as it may introduce unexpected behavior in the user's code and also
provides no way to attach an error handler besides manually wrapping
the bugsnag library.

PHPUnit's default behavior is to translate PHP errors into exceptions
and raise those. Rely on this assumption in our tests of the error
handler.

Ref: https://phpunit.de/manual/current/en/writing-tests-for-phpunit.html

* Switch to backwards compatible PHPUnit annotation

Since this library needs to be compatible with older versions of PHP,
CI tests run on two versions of PHPUnit.

4.8 doesn't have PHPUnit_Framework_TestCase::expectException, however
both newer and older versions of PHPUnit provide support for the
annotation @expectedException, so use that instead.

* Add test to check custom error handler value is returned

* Return value from previousErrorHandler

* Update phpdoc

* Preserve register behavior; add registerWithPrevious

* Update testCustomErrorHandlerValueReturned

This test should have failed when changing the default behavior not to
call the previous error handler. Instead of returning false, which is
the default from the Bugsnag handler, and thus proves nothing, return a
sentinel value and assert based on that.

* $callPrevious should not be an optional parameter

Since these are new methods expected to be invoked primarily by register
or registerWithPrevious, make the boolean flag a required parameter.

A user might wish to only register certain handlers themselves, so it
makes sense to leave these public, but the behavior is different
depending on the flag. The caller should be forced to make an explicit
decision.

* Update docblocks