Use HttpClient in SessionTracker
This change unifies the SessionTracker and Client WRT how they use
Guzzle. They now both use the HttpClient abstraction, which also
allows them to share a Guzzle instance. This means that we can't
use the base URI feature of Guzzle anymore because the notify and
session endpoints are on separate subdomains. However this isn't a
big deal as we POST to the root in most cases anyway; the only
exception being the deploy notification, which is '/deploy'

BC breaks:
- Removal of Client and Configuration 'getSessionClient'
  This doesn't make sense to keep given the session client is now
  the same as the notify Guzzle client. Keeping this would mean
  changes to the "session" client also propagate to the notify
  client, which could lead to things being delivered to the wrong
  endpoint

- We no longer use the Guzzle base URI for requests
  This means if a Guzzle client is passed in the constructor with
  a pre-defined base URI, we will still send requests to the notify
  URI. Client::setNotifyEndpoint now needs to be called manually
  instead. This changed because the Guzzle base URI is ambiguous now
  given that it is shared between notifications and sessions

- Removal of SessionTracker::setConfig
  This was unused internally and doesn't seem to be needed as the
  configuration can be changed via the Client or Configuration
  instance itself. Having the posibility for there to be two different
  Configuration instances could be dangerous

Additionally several constants have been deprecated as they are no
longer used. Specifically these are:

- Client::ENDPOINT
  This is ambiguous as we have three separate endpoints
  Use Configuration::NOTIFY_ENDPOINT instead

- HttpClient::PAYLOAD_VERSION
  This is ambiguous as there is a session payload version too
  Use HttpClient::NOTIFICATION_PAYLOAD_VERSION instead

- Report::PAYLOAD_VERSION
  As above. This was also unused by the notifier
  Use HttpClient::NOTIFICATION_PAYLOAD_VERSION instead

- SessionTracker::$SESSION_PAYLOAD_VERSION
  Use HttpClient::SESSION_PAYLOAD_VERSION instead

Finally, Client::makeGuzzle has been deprecated as it is only used
in one place internally to Client and therefore doesn't need to be
public