Allow the current session to start mutliple times
This is necessary to allow batching of sessions in the SessionTracker

If we can only start a session once, then we can only have a single
session in the SessionTracker too. However, this doesn't work because
SessionTracker has a 'startSession' method which can be called
multiple times. Therefore we need to support starting the current
session as many times as 'SessionTracker::startSession' is called