Batch sessions to be sent on destruction
This solves a problem with the previous implementation, where it was
possible to lose sessions

The first time 'startSession' is called it would immediately send
sessions as 'lastSent' is initialised to '0'. The second time it's
called, 'lastSent' will be too recent to send the second session.
If 'startSession' or 'sendSessions' are not called after this, the
second session would be lost completely

With this commit we have no concept of a 'lastSent' time; instead
'startSession' only adds the new session to the SessionTracker's
internal session count list. Then, on destruct, we send the entire
list at once. This ensures we only ever send one HTTP request
unless 'sendSessions' is manually called

It is still possible to lose sessions if 'sendSessions' fails and
isn't retried, but that is always going to be possible — HTTP can
and will fail and retrying infinitely doesn't guarantee success