Avoid using 'setUp' and 'tearDown' method in tests
The replacements use '@before' & '@after' hooks, which are supported
on all PHPUnit versions we support. The name of the methods aren't
important anymore, but are the same for consistency

We have to do this because PHPUnit 9 has added 'void' return type
declarations to its abstract TestCase class. This means we have to
have the same type declaration in our tests. We can't do that though,
because we still support PHP 5 and return type declarations weren't
added until PHP 7